fix #4: "licenses" attribute name collision
My license report functionality was in an attribute called "licenses",
which overrode the normal `lib.licenses` atttribute in Nixpkgs that has
all the licenses.

This fix is merely moves the name from "licenses" to "license-report".

The reason I didn't catch this error is because I had
`allowUnfree = true;` in my `~/.nixpkgs/config.nix`.  This made the
broken lookup of `lib.licenses.bsd3` not throw an error.

On any Nix installation without this allowance, the build broke (as it
should have).

The lesson I learned is... if you want deterministic builds in Nix, be
extremely careful of what you put in `~/.nixpkg/config.nix`.