fix: extn-haskell Dante support
Dante changed their API recently to have a
dante-repl-command-line-methods in addition to the alist of functions.

This is similar to the extn-haskell/dante-repl-types variable, but at
Dante's level.

The problem was that functions were being looked up in the alist based
on how this new dante-repl-command-line-methods varaible was set.  So
this change just sets it to the same setting as
extn-haskell/dante-repl-types.