spacemacs: remove floobits
This Spacemacs configuration is really my personal configuration.  On
one hand, it's not a bad one -- it's fairly minimal.  But sometimes I'm
trying out a tool that I couldn't necessarily recommend.  In this case,
it was Floobits.

Maybe I'll figure out a better way to split out my spacemacs
configuration into a published configuration and an unpublished overlay.

For now, I like the idea that I'm actually using the configuration I'm
sharing with others.  It's as close as I get to testing.