docs: regenerate all Markdown + one small typo fix
Unfortunately, I've switched computers from Debian to NixOS, and for
some reason, the Emacs packages pulled in and set up don't have the
latest GitHub Flavored Markdown generation.  So the generation of
anchors for the TOC is no longer deterministic.  Kind of a bummer, would
be nice to figure out, but it's not like it's breaking anything.