cabal-new-watch: now works on Macs
I developed this tool on a Linux box, and didn't think about inotifywait
not being available on Macs.  I switched to the fswatch tool, which has
multiple backends to over both Linux and Mac systems.

I also improved the script a little addressing shellcheck warnings and
making it look at the md5sum of the triggering file to avoid some (maybe
not all) needless rebuilds.