cabal-new-watch: fix: works on Mac now
In a previous commit, I presumed that if I could get fswatch working on
a Linux machine, it would work similarly enough on a Mac.  That was a
wrong assumption.  When I finally got a Mac, I found out that different
events are emitted by different backends.  So I found a set of events
that seems to work well enough both for Linux and for Mac.

I also removed the "-d" switch, which I guess didn't do what I thought
it did.  It was benign in Linux (seems to have done nothing), but on
Macs, it only gave events for files.

This change has been tested on both a Linux and Mac machine, so I have
some confidence in it.  But the reality is that I don't use this script
that much in practice (I rely more on Dante in Emacs).  So I welcome
improvements if people have ideas.