Update Nixpkgs revision, and cascading changes
I wanted to update the Nixpkg version, but this also led me to want to
update the default GHC from 8.2.2 to 8.4.3 (8.6.1 has a [horrible
defect] still).

Fortunately, in the 8.4.3 series of compilation on GitHub, the tests for
Hasktags are now passing, so I can stop overriding that package to
disable tests.

When running tests, the code for dockerTools upstream seemed to change
from defaulting image tags to "lastest" to something more explicit.
This is great from a safety perspective, but it broke some scripts, so I
just did the simple fix of explicitly creating "lastest"-tagged images.
in the Nix expression.

Changing the build.nix files unfortunately creates some shifts of line
numbers in the documentation, so I updated that.

And finally, the biggest change... I switched from defaulting Pkgs-make
to the "curated" Nixpkgs overlay to the no-op ("plain") one.  With the
Hasktags tests fixed, the only thing left curated are some Python
machine-learning packages.  But honestly, I can't see a convenient path
to keep this curation moving forward.  I'm thinking of deleting this
from the repository all-together.  I changed the documentation not call
out this curation so much, so people have to dig deep to find the one
mention of it, and even then it's clear that this curation is
experimental.

[horrible defect]: https://ghc.haskell.org/trac/ghc/ticket/15696