pkgs-make: haskell: option to force cabal2nix
There might be a default.nix in a directory, but we want to ignore it
and generate a Nix expression from cabal2nix.

This came up when building a single-module project with Pkgs-make.  If
we don't put the Cabal project in its own subdirectory, we may want the
default.nix file to be for the top-level build, not the callPackage Nix
expression from cabal2nix.