spacemacs: dotfile: "codex update" safe command
Projectile doesn't make the `projectile-tags-command` safe by default
because that could be a Trojan horse for arbitrary scripts to be
injected with .dir-locals.el files.  So we have to declare explicitly
safe settings for it.

At work, I started using Codex for Haskell tags generation because it's
faster than Haskdogs.  This may influence some changes to the
nix-tags-haskell script.