spacemacs: haskell: have Dante ignore .stack files
Dante tries to start up a REPL for all files, but that's annoying for
source without a project, which is often the case for source cached
dotfiles.  We use this source often as references for tag-generation.

This commit just adds in .stack as one of the ignored directories.  I
started used Stack at work, so this came up, but it's not a bad setting
by default.