spacemacs: dotfile: disable floating tooltip
The tooltip very often covers up the code it's referring to.  And on top
of that, it disappears before it can be read when there's a lot of
information in it.

This is a feature that only works for small messages, and is ill-suited
for compiler errors/warnings, which is what people use it for often.