pkgs-make: remove curated packages
This was a hold-over from an era when I wanted to maintain a bunch of
machine learning Nix expressions.  But that work is just larger than I
can take on, and I don't have a good test bed.  It's just a hard
problem.

Curation instead will now just become topical fixes.  Sometimes Nixpkgs
is updated to be current, but needs a few fixes.