spacemacs: haskell: catch up with upstream changes
The standard `haskell` layer that `extn-haskell` depends on had improved
to include both `attrap` and `dante` packages.  This commit streamlines
`extn-haskell` and fixes a defect with respect to removing hooks (the
name of the function changed).  This is why the ability to not flycheck
files by regex stopped working.  We weren't unregistering the hook that
had the default logic.